DEA Flask App/requirements.txt
flask
werkzeug==3.0.4
psycopg2-binary==2.9.9
jsonify
flask-cors==4.0.1
cryptography
